refactor: cross-feature import 위반 해소 + event_now_bottom_sheet 5-phase 분리
home→event/ticket 위반: MatchingVoteContent, TicketQRViewer를 common/widgets/로 이동.
eventTicketTokenProvider를 common/으로 이동.
event_now_bottom_sheet.dart를 5개 phase 파일로 분리 (871줄→98줄).

event→home 위반: goToPurchaseHistory()를 EventCoordinator로 이동.
event→ticket 위반: showTicketSelection()을 EventCoordinator로 통합.

Closes #1094, closes #1096

Constraint: TicketSelectionSheet는 ticket feature 내부 의존이 깊어 common 이동 불가
Rejected: TicketSelectionSheet를 common으로 이동 | ticket_recommendation_util 등 내부 의존이 너무 많음
Directive: EventCoordinator의 ticket import는 coordinator seam으로 의도적 허용 — widget/page에서 직접 참조 금지
Confidence: high
Scope-risk: moderate
Not-tested: TicketSelectionSheet 표시 후 onTicketSelected 콜백 동작
