+++ apps/app_user/lib/src/features/event/logic/event_coordinator.dart
@@ -1,7 +1,9 @@
 import 'dart:async';
 
+import 'package:app_user/src/features/ticket/ui/ticket_selection_sheet.dart';
 import 'package:app_user/src/routing/app_router.dart';
 import 'package:app_user/src/routing/app_routes.dart';
+import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:minglit_kit/minglit_kit.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
@@ -48,5 +50,28 @@ class EventCoordinator {
   // Fix #634: auth_coordinator 직접 참조 → event_coordinator 내부로 login 네비게이션 위임
   void pushLogin({String? from}) {
     unawaited(_router.push(LoginRoute(from: from).location));
+  }
+
+  // Fix #1094: home_coordinator 직접 참조 제거 — event feature 내에서 구매내역 네비게이션
+  void goToPurchaseHistory() {
+    _router.go(const PurchaseHistoryRoute().location);
+  }
+
+  // Fix #1094: ticket_coordinator 직접 참조 제거 — event feature 내에서 티켓 선택 UI 관리
+  void showTicketSelection(
+    BuildContext context,
+    Event event, {
+    required void Function(String eventId, String? ticketId) onTicketSelected,
+  }) {
+    unawaited(
+      showModalBottomSheet<void>(
+        context: context,
+        isScrollControlled: true,
+        builder: (_) => TicketSelectionSheet(
+          event: event,
+          onTicketSelected: onTicketSelected,
+        ),
+      ),
+    );
   }
 }
